feature: add steps in reminder creation
add step in reminder creation
adapt edit reminder view
